requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
celery
redis
onnxruntime
structlog
apscheduler
